fix(build): repair C1f split fallout so the full tree compiles and links
Upstream 0089022/78ba153 (C1f backend split) and the O1.x/binding-domain
refactors broke several mixed C/C++/ObjC TUs; only the new TUs were
smoke-built on Linux. Local repairs, none of which change GL behavior:

- mgl_air_backend.cpp: restore samplerTexValue/scalarFromType helpers
  and alias mgl::air::callAirFn (definition moved into mgl_air_matrix.cpp
  without leaving a declaration); declare callAirFn and include
  ArrayRef.h in the matrix header
- mgl_types_sync.h / mgl_frame_activity.h: use <atomic> under C++ and
  <stdatomic.h> under C via #ifdef __cplusplus + MGL_ATOMIC(T) macro
  (libc++ <stdatomic.h>/<atomic> sentinel clash; C++23 unusable with
  LLVM 15 APFloat)
- batch/draw encode ports (6 .m files): hoist @implementation above the
  file-scope C entry points, replace MGL_STATE() (ObjC-only dual proxy)
  with ctx->active_state inside C functions, resolve the device via
  mglRendererBackendGetDevice instead of ivar macros, cast narrowing
  exprs; add MGLRenderer+BatchPorts_Private.h for cross-port selectors
- Makefile: keep -std=c++20, compile mgl_ir.c to a separate object for
  test_mgl_air_type (C++ std cannot apply to C input), add -isysroot to
  the two binding test rules
- gen_aux_assets.py: wrap generated declarations in extern "C" for the
  C++ consumer; regenerate mgl_aux_assets.{h,c}
- misc: include-order fixes (glm_dispatch/glm_context before
  draw_command), named struct for a C++ cast (mgl_draw_cull.cpp),
  comment terminator fix (MGLRenderer+Draw_Private.h), tautological
  bitwise fix in test_xfb_plan.c

Verified with a full clean rebuild + make test-all. Two upstream
regressions remain and are diagnosed but intentionally not fixed here:
test-regression air_geometry_resources SIGSEGV (ca96541 truncates GS
compute-plan temporaries lifetime: temps die in mglGsMetalFillComputeBindings
before the plan is encoded) and test-mglair TCS factor zeros.

+++ MGL/src/mgl_batch_mtl_encode.cpp
@@ -11,6 +11,9 @@
 #include "mgl_index_buffer.h"
 
 #include "mgl_batch_restore.h"
+/* draw_command.h expects GL base types (GLenum/GLuint/...) to already be
+ * declared; pull in the canonical dispatch chain first. */
+#include "glm_dispatch.h"
 #include "draw_command.h"
 #include "mgl_render.h"
 
@@ -501,9 +504,9 @@ extern "C" int mgl_batch_mtl_issue_stream_mdi_batch(
     }
     size_t neededBytesRaw = 0;
     MGLBatchStreamMdiGateIn gateIn = {
-        .stream_merged = batch->stream_merged ? 1u : 0u,
-        .has_encoder = render_encoder_owner ? 1u : 0u,
-        .disable_mdi = disable_mdi ? 1u : 0u,
+        .stream_merged = (uint8_t)(batch->stream_merged ? 1u : 0u),
+        .has_encoder = (uint8_t)(render_encoder_owner ? 1u : 0u),
+        .disable_mdi = (uint8_t)(disable_mdi ? 1u : 0u),
         .primitive_type = batch->key.primitive_type,
         .command_count = batch->command_count,
         .stream_index_count = (uint32_t)batch->stream_index_count,
